+++ drake/BUILD
@@ -56,17 +56,16 @@ _LIBDRAKE_COMPONENTS = [
     "//drake/common:functional_form",
     "//drake/common:is_approx_equal_abstol",
     "//drake/common:is_cloneable",
-    "//drake/common:matlab_rpc",
     "//drake/common:monomial",
     "//drake/common:nice_type_name",
     "//drake/common:number_traits",
     "//drake/common:polynomial",
-    "//drake/common:protobuf",
     "//drake/common:reinit_after_move",
     "//drake/common:scoped_singleton",
     "//drake/common:sorted_vectors_have_intersection",
     "//drake/common:symbolic",
     "//drake/common:text_logging_gflags",
+    "//drake/common/proto:protobuf",
     "//drake/common/trajectories:piecewise_function",
     "//drake/common/trajectories:piecewise_polynomial",
     "//drake/common/trajectories:piecewise_polynomial_trajectory",
