+++ drake/common/proto/test/test_string.txt
