+++ drake/examples/Acrobot/BUILD
@@ -111,8 +111,8 @@ drake_cc_binary(
     test_rule_args = ["-realtime_factor=0.0"],
     deps = [
         ":acrobot_plant",
-        "//drake/common:call_matlab",
         "//drake/common:find_resource",
+        "//drake/common/proto:call_matlab",
         "//drake/lcm",
         "//drake/multibody:rigid_body_tree",
         "//drake/multibody/joints",
