+++ drake/examples/QPInverseDynamicsForHumanoids/param_parsers/BUILD
@@ -21,7 +21,7 @@ drake_cc_library(
     deps = [
         ":id_controller_config",
         ":rigid_body_tree_alias_groups",
-        "//drake/common:protobuf",
+        "//drake/common/proto:protobuf",
         "//drake/examples/QPInverseDynamicsForHumanoids:qp_controller_common",
         "@yaml_cpp",
     ],
@@ -47,7 +47,7 @@ drake_cc_library(
     hdrs = ["rigid_body_tree_alias_groups.h"],
     deps = [
         ":alias_groups",
-        "//drake/common:protobuf",
+        "//drake/common/proto:protobuf",
         "//drake/multibody:rigid_body_tree",
     ],
 )
