+++ drake/examples/QPInverseDynamicsForHumanoids/param_parsers/param_parser.cc
@@ -3,7 +3,7 @@
 #include "google/protobuf/text_format.h"
 
 #include "drake/common/drake_assert.h"
-#include "drake/common/protobuf.h"
+#include "drake/common/proto/protobuf.h"
 #include "drake/examples/QPInverseDynamicsForHumanoids/param_parsers/id_controller_config.pb.h"
 #include "drake/util/drakeUtil.h"
