+++ drake/solvers/BUILD
@@ -241,7 +241,7 @@ drake_cc_library(
     ],
     deps = [
         ":rotation_constraint",
-        "//drake/common:call_matlab",
+        "//drake/common/proto:call_matlab",
     ],
 )
 
@@ -806,7 +806,7 @@ drake_cc_binary(
     deps = [
         ":mathematical_program",
         ":rotation_constraint",
-        "//drake/common:call_matlab",
+        "//drake/common/proto:call_matlab",
     ],
 )
