+++ drake/solvers/test/rotation_constraint_visualization.cc
@@ -1,6 +1,6 @@
 #include "drake/solvers/test/rotation_constraint_visualization.h"
 
-#include "drake/common/call_matlab.h"
+#include "drake/common/proto/call_matlab.h"
 #include "drake/solvers/rotation_constraint.h"
 #include "drake/solvers/rotation_constraint_internal.h"
